bot/requirements.txt
aiogram>=3.0.0
aiohttp>=3.8.0
python-dotenv>=1.0.0
pydantic>=2.0.0
